+++ src/routes/register/+page.svelte
@@ -7,8 +7,8 @@
 	import Footer from 'elements/navigation/footer.svelte';
 
 	import Spacer from 'elements/spacer/spacer.svelte';
-	import MessageWrapper from 'elements/messages/messageWrapper.svelte';
-	import ErrorMessage from 'elements/messages/errorMessage.svelte';
+	import MessageWrapper from 'elements/message/messageWrapper.svelte';
+	import ErrorMessage from 'elements/message/errorMessage.svelte';
 	import Input from 'elements/input/input.svelte';
 	import TextLine from 'elements/text/textLine.svelte';
 	import List from 'elements/list/list.svelte';
